requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.9.0
django-environ==0.7.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.4
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
